requirements.txt

fastapi==0.115.6
python-dotenv==1.2.2
groq==1.1.1
sentence-transformers==5.3.0
joblib==1.3.2
pandas==2.0.2
scikit-learn==1.6.0
